feat: show IndemniteLegale for conventions without indemnite
